+++ IDisposableAnalyzers.Test/Helpers/DisposableMemberTests.cs
@@ -70,6 +70,33 @@ namespace N
             Assert.AreEqual(Result.Yes, DisposableMember.IsDisposed(new FieldOrPropertyAndDeclaration(symbol, declaration), semanticModel, CancellationToken.None));
         }
 
+        [Test]
+        public static void DisposedInExplicitImplementation()
+        {
+            var syntaxTree = CSharpSyntaxTree.ParseText(@"
+namespace N
+{
+    using System;
+    using System.IO;
+
+    sealed class C : IDisposable
+    {
+        private readonly IDisposable stream = File.OpenRead(string.Empty);
+
+        void IDisposable.Dispose()
+        {
+            this.stream.Dispose();
+        }
+    }
+}");
+            var compilation = CSharpCompilation.Create("test", new[] { syntaxTree }, MetadataReferences.FromAttributes());
+            var semanticModel = compilation.GetSemanticModel(syntaxTree);
+            var declaration = syntaxTree.FindFieldDeclaration("stream");
+            var symbol = semanticModel.GetDeclaredSymbolSafe(declaration, CancellationToken.None);
+            var field = new FieldOrPropertyAndDeclaration(symbol, declaration);
+            Assert.AreEqual(Result.Yes, DisposableMember.IsDisposed(field, semanticModel, CancellationToken.None));
+        }
+
         [TestCase("this.components.Add(this.stream)")]
         [TestCase("components.Add(stream)")]
         public static void FieldAddedToFormComponents(string expression)