+++ ValidCode.NetCore/ExplictImplementaion.cs
@@ -1,0 +1,15 @@
+﻿namespace ValidCode.NetCore
+{
+    using System;
+
+    public sealed class ExplictImplementaion : IDisposable
+    {
+        private Disposable? disposable = new Disposable();
+
+        void IDisposable.Dispose()
+        {
+            this.disposable?.Dispose();
+            this.disposable = null;
+        }
+    }
+}